refactor: repostory 위치 변경
