+++ app/src/main/java/woowacourse/shopping/data/recentproduct/RecentProductLocalRepository.kt
@@ -3,11 +3,10 @@ package woowacourse.shopping.data.recentproduct
 import android.content.Context
 import com.example.domain.product.Product
 import com.example.domain.product.recent.RecentProduct
-import com.example.domain.product.recent.RecentProductRepository
 import woowacourse.shopping.util.BANDAL
 import java.time.LocalDateTime
 
-class RecentProductRepositoryImpl(
+class RecentProductLocalRepository(
     context: Context,
     url: String,
     user: String = BANDAL
