+++ app/src/main/java/woowacourse/shopping/feature/product/detail/ProductDetailPresenter.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.feature.product.detail
 
-import com.example.domain.cart.CartRepository
+import woowacourse.shopping.data.cart.CartRepository
 import woowacourse.shopping.databinding.DialogSelectCountBinding
 import woowacourse.shopping.model.CartProductState.Companion.MAX_COUNT_VALUE
 import woowacourse.shopping.model.CartProductState.Companion.MIN_COUNT_VALUE
