+++ app/src/main/java/woowacourse/shopping/util/Injector.kt
@@ -1,14 +1,14 @@
 package woowacourse.shopping.util
 
 import android.content.Context
-import com.example.domain.cart.CartRepository
-import com.example.domain.product.ProductRepository
-import com.example.domain.product.recent.RecentProductRepository
 import retrofit2.Retrofit
 import woowacourse.shopping.data.cart.CartRemoteRepository
+import woowacourse.shopping.data.cart.CartRepository
 import woowacourse.shopping.data.order.OrderRemoteRepository
 import woowacourse.shopping.data.product.ProductRemoteRepository
-import woowacourse.shopping.data.recentproduct.RecentProductRepositoryImpl
+import woowacourse.shopping.data.product.ProductRepository
+import woowacourse.shopping.data.recentproduct.RecentProductLocalRepository
+import woowacourse.shopping.data.recentproduct.RecentProductRepository
 import woowacourse.shopping.data.util.RetrofitManager
 
 object Injector {
@@ -27,7 +27,7 @@ object Injector {
 
     fun provideRecentProductRepository(context: Context, serverUrl: String): RecentProductRepository {
         if (this.serverUrl == null) this.serverUrl = serverUrl
-        return RecentProductRepositoryImpl(context, serverUrl)
+        return RecentProductLocalRepository(context, serverUrl)
     }
 
     fun provideOrderRemoteRepository(serverUrl: String): OrderRemoteRepository {
